+++ WebContent/emp/register.jsp
@@ -5,29 +5,55 @@
 <head>
 <meta charset="EUC-KR">
 <title>Insert title here</title>
- <link rel="stylesheet" href="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css">
-  <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.0/jquery.min.js"></script>
-  <script src="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"></script>
-</head>
+ </head>
 <body>
 	<div class="container-fluid">
 			<form role="form" class="form-horizontal"
-				action="/serverweb/register.do" method="POST" 
+				action="/serverweb/emp/insert.do" method="POST" 
 				name="myform">
 				<fieldset>
 					<div id="legend">
 						<legend>아래 양식을 작성해주세요.</legend>
 					</div>
+					<div class="form-group" style="padding: 40px">
+						<p class="centered">
+							<img src="/kimsaemERP/images/myphoto.jpg" 
+							 id="userImage" style="width: 100px">
+						</p>
+						<div>
+							<input type="file" name="userImage"
+								onchange="document.getElementById('userImage').src = window.URL.createObjectURL(this.files[0])"
+								accept="image/*">
+						</div>
+					</div>
+					
 					<div class="form-group">
 						<!-- 부서코드 -->
 						<label class="control-label col-sm-2" for="orgcode">부서코드</label>
 						<div class="col-sm-3">
-							<input type="text" id="orgcode" name="deptno"
-								placeholder="부서코드" class="form-control"
-								 required>
+							<select name="deptno" class="form-control" >
+								<option value="d001">인사팀
+								<option value="d002">재무팀
+								<option value="d003">개발팀
+								<option value="d0031">디자인팀
+								<option value="d0032">Ajax팀
+								<option value="d004">개발지원
+								<option value="d005">시스템지원팀
+								<option value="d006">총무과
+								<option value="dg001">경영지원본부
+								<option value="dg002">영업본부
+								<option value="dg003">IT지원센터
+								<option value="d007">교육부
+								<option value="d008">기획실
+								<option value="d009">영업1팀
+								<option value="d010">기업영업본부
+								<option value="d011">영업2팀
+								<option value="d012">마케팅실
+								
+							</select>
 						</div>
 					</div>
-
+		
 					
 					
 					<div class="form-group">
@@ -50,48 +76,109 @@
 						</div>
 						<span id="checkVal"></span>
 					</div>
-
-
-
-				
-
-
-					
 					<div class="form-group">
 						<!-- 패스워드-->
 						<label class="control-label col-sm-2" for="pass">패스워드</label>
 						<div class="col-sm-3">
 							<input type="text" id="pass" name="pass"
-								placeholder="패스워드" class="form-control" minlength="4" >
+								placeholder="패스워드" class="form-control" minlength="15" >
 
+						</div>
+					</div>
+					<div class="form-group">
+						<!-- 주민번호-->
+						<label class="control-label col-sm-2" for="ssn">주민번호</label>
+						<div class="col-sm-3">
+							<input type="text" id="ssn" name="ssn"
+								placeholder="주민번호" class="form-control" 
+								minlength="15" >
+
+						</div>
+					</div>
+					<div class="form-group">
+						<!-- 생일-->
+						<label class="control-label col-sm-2" for="birthday">생년월일</label>
+						<div class="col-sm-3">
+							<input type="text" id="birthday" name="birthday"
+								placeholder="생년월일" class="form-control" 
+								minlength="15" >
+
+						</div>
+					</div>
+					<div class="form-group">
+						<!-- 결혼유무-->
+						<label class="control-label col-sm-2" for="marry">결혼유무</label>
+						<div class="col-sm-3">
+							<input type="checkbox" id="marry" name="marry"
+								placeholder="결혼유무"	minlength="15" >결혼유무
+
+						</div>
+					</div>
+					<div class="form-group">
+						<label class="col-sm-2 col-sm-2 control-label">우편번호</label>
+						<div class="col-sm-3">
+							<input type="text" class="form-control" name="zipcode"
+								id="zipcode" >
+						</div>
+						<div class="col-sm-2">
+							<button type="button"
+								class="btn btn-round btn-primary form-control">우편번호 검색</button>
 						</div>
 					</div>
 					<div class="form-group">
 						<!-- 주소-->
 						<label class="control-label col-sm-2" for="addr">주소</label>
-						<div class="col-sm-3">
+						<div class="col-sm-6">
 							<input type="text" id="addr" name="addr" 
 							placeholder="주소"
 								class="form-control" minlength="4" required>
+							
+						</div>
+					</div>
+					<div class="form-group">
+						<label class="col-sm-2 col-sm-2 control-label"></label>
+						<div class="col-sm-5">
+							<input type="text" class="form-control" name="detailaddr"
+								id="detailaddr" >
+						</div>
+					
+					</div>
+					<div class="form-group">
+						<!-- 집전화-->
+						<label class="control-label col-sm-2" for="phonehome">집전화</label>
+						<div class="col-sm-5">
+							<input type="text" id="phonehome" name="phonehome" 
+							placeholder="집전화"
+								class="form-control" minlength="15" required>
 
 						</div>
 					</div>
 					<div class="form-group">
-						<!-- 포인트-->
-						<label class="control-label col-sm-2" for="point">포인트</label>
-						<div class="col-sm-3">
-							<input type="text" id="point" name="point" 
-							placeholder="포인트"
-								class="form-control" minlength="4" required>
+						<!-- 회사전화-->
+						<label class="control-label col-sm-2" for="phoneco">회사전화</label>
+						<div class="col-sm-5">
+							<input type="text" id="phoneco" name="phoneco" 
+							placeholder="회사전화"
+								class="form-control" minlength="15" required>
+
+						</div>
+					</div>
+					<div class="form-group">
+						<!-- 집전화-->
+						<label class="control-label col-sm-2" for="phonehome">핸드폰번호</label>
+						<div class="col-sm-5">
+							<input type="text" id="phonecell" name="phonecell" 
+							placeholder="핸드폰번호"
+								class="form-control" minlength="15" required>
 
 						</div>
 					</div>
 					<div class="form-group">
 						<!-- 등급-->
-						<label class="control-label col-sm-2" for="grade">등급</label>
+						<label class="control-label col-sm-2" for="grade">이메일</label>
 						<div class="col-sm-3">
-							<input type="text" id="grade" name="grade" 
-							placeholder="등급"
+							<input type="text" id="email" name="email" 
+							placeholder="이메일"
 								class="form-control" minlength="4" required>
 
 						</div>
@@ -105,6 +192,6 @@
 				</fieldset>
 			</form>
 	</div>
-	<div><a href="/serverweb/list.do">회원목록보기</a></div>
+	<div><a href="/serverweb/emp/list.do">회원목록보기</a></div>
 </body>
 </html>
