+++ WebContent/menu/insa_menu.jsp
@@ -8,10 +8,10 @@
 </head>
 <body>
 	<div class="list-group">
-		<a href="#" class="list-group-item">부서등록</a> 
-		<a href="#"	class="list-group-item">인사등록</a> 
+		<a href="/kimsaemERP/view.html?menupath=/menu/insa_menu.jsp&viewpath=/dept/dept_register.jsp" class="list-group-item">부서등록</a> 
+		<a href="/kimsaemERP/view.html?menupath=/menu/insa_menu.jsp&viewpath=/emp/register.jsp"	class="list-group-item">인사등록</a> 
 		<a href="#" class="list-group-item">부서조회</a>
-		<a href="#" class="list-group-item">부서별 인사 조회</a>
+		<a href="/kimsaemERP/view.html?menupath=/menu/insa_menu.jsp&viewpath=/emp/emp_list.jsp" class="list-group-item">부서별 인사 조회</a>
 		<a href="#"	class="list-group-item">인사평가</a>
 		<a href="#" class="list-group-item">인사점수조회하기</a>
 	</div>
