+++ src/erp/controller/logoutServlet.java
@@ -1,4 +1,4 @@
-package emp.controller;
+package erp.controller;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
